ADEN-2044 moved Mercury test to the same file where Oasis tests are
