+++ src/test/java/com/wikia/webdriver/testsuites/testSuite.xml
@@ -24,7 +24,6 @@
             <class name="com.wikia.webdriver.testcases.adstests.TestAdsOnSpecial"/>
             <class name="com.wikia.webdriver.testcases.adstests.TestAdsScreenshotComparison"/>
             <class name="com.wikia.webdriver.testcases.adstests.TestAmazonAds"/>
-            <class name="com.wikia.webdriver.testcases.adstests.TestAmazonMobileAds"/>
             <class name="com.wikia.webdriver.testcases.adstests.TestATF"/>
             <class name="com.wikia.webdriver.testcases.adstests.TestDfpParamsPresent"/>
             <class name="com.wikia.webdriver.testcases.adstests.TestDisableGptAds"/>
